♻️ refactor: RefreshToken 코드 다 삭제

+++ src/main/java/com/uahannam/member/entity/Role.java
@@ -12,7 +12,7 @@ public class Role extends SysTimeCols{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private Integer roleId;
+    private Long roleId;
 
     @Column(name = "role_name")
     private String roleName;
